backend/food-detector/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy

Pillow
python-multipart

tensorflow
Orange3
Orange3-ImageAnalytics
pymongo